Enables dynamic content updates from CMS
Introduces a mechanism to fetch and apply content updates from a CMS to the storefront dynamically.

This includes adding GitHub client to handle file creation/updates, and adjusting the deploy middleware to use it for committing generated JSON files.

Removes unused commands and environment configurations.
Adds outbound access to Github API in manifest.