Adds multiple static banner section
Adds a new section type to the homepage builder for displaying multiple static banners.

This section allows content managers to create rows of banners, each with its own image, mobile image, and link. The component also allows to configure the number of banners to display per row on different devices and adds configuration of the gap between banners.